requirements.txt
numpy
joblib
pydantic
scikit-learn==1.6.1
google-adk
